C:\Users\zainu\Desktop\курс\1proekt\1proekt\Ozhaev Zaynudin cartoon.cpp -o C:\Users\zainu\Desktop\курс\1proekt\1proekt\Ozhaev Zaynudin cartoon.exe  -Wshadow -Winit-self -Wredundant-decls -Wcast-align -Wundef -Wfloat-equal -Winline -Wunreachable-code -Wmissing-declarations -Wmissing-include-dirs -Wswitch-enum -Wswitch-default -Weffc++ -Wmain -Wextra -Wall -g -pipe -fexceptions -Wcast-qual -Wconversion -Wctor-dtor-privacy -Wempty-body -Wformat-security -Wformat=2 -Wignored-qualifiers -Wlogical-op -Wmissing-field-initializers -Wnon-virtual-dtor -Woverloaded-virtual -Wpointer-arith -Wsign-promo -Wstack-usage=8192 -Wstrict-aliasing -Wstrict-null-sentinel -Wtype-limits -Wwrite-strings -D_DEBUG -D_EJUDGE_CLIENT_SIDE -D_EJC

+++ Ozhaev Zaynudin cartoon.cpp
@@ -13,19 +13,19 @@
 
 void DrawMir        ();
 void DrawZoomHouse1 ();
-void DraTrainFare   (int t);
-void DrawTrain      (int x,     int y);
-void DrawHouse2     (int x,     int y);
-void DrawHouse3     (int x,     int y);
-void DrawTrain      (int x,     int y, double sizeX, double sizeY);
-void DrawTree1      (int x,     int y, double sizeX, double sizeY, COLORREF TreeColor);
-void DrawTree2      (int x,     int y, double sizeX, double sizeY, COLORREF TreeColor);
-void DrawSun        (int x,  double y, double sizeX, double sizeY, COLORREF SunColor);
-void DrawGirl       (int x,     int y, double sizeX, double sizeY, COLORREF GirlColor,
-                                       double eyes,  double smayl);
-void DrawMan        (int x,     int y, double sizeX, double sizeY, COLORREF ManColor,
-                                       double eyes,  double smayl);
-void DrawHouse1     (double x,     double y, double sizeX, double sizeY, COLORREF HoleColor);
+void DraTrainFare   (   int t);
+void DrawTrain      (   int x,     int y);
+void DrawHouse2     (   int x,     int y);
+void DrawHouse3     (   int x,     int y);
+void DrawTrain      (   int x,     int y, double sizeX, double sizeY);
+void DrawTree1      (   int x,     int y, double sizeX, double sizeY, COLORREF TreeColor);
+void DrawTree2      (   int x,     int y, double sizeX, double sizeY, COLORREF TreeColor);
+void DrawSun        (   int x,  double y, double sizeX, double sizeY, COLORREF SunColor);
+void DrawGirl       (   int x,     int y, double sizeX, double sizeY, COLORREF GirlColor,
+                                       double eyes , double smayl);
+void DrawMan        (   int x,     int y, double sizeX, double sizeY, COLORREF ManColor,
+                                       double eyes , double smayl);
+void DrawHouse1     (double x,  double y, double sizeX, double sizeY, COLORREF HoleColor);
 void DrawBackground (COLORREF SkyColor);
 
 int main()
@@ -368,8 +368,8 @@ void DrawZoomHouse1()
                 txClear ();
                 DrawBackground (RGB (0, 80, 80));
                 DrawHouse1 (320 + t*14.2, 260 + t*22.5 , 1 + t*0.24, 1 + t*0.2, RGB (0, 80, 80));
-                int x = ROUND( 260 - t*2.6 );
-                int y = 163 - t*1.7;
+                double x = ROUND( 260 - t*2.6 );
+                double y = 163 - t*1.7;
                 double sizeX = 0.006*t;
                 double sizeY = 0.0037*t;
 
@@ -571,8 +571,8 @@ void DrawZoomHouse1()
                 }
             else
                 {
-                int x = ROUND( 260 - t*2.6 );
-                int y = 163 - t*1.7;
+                double x = ROUND( 260 - t*2.6 );
+                double y = 163 - t*1.7;
                 double sizeX = 0.006*t;
                 double sizeY = 0.0037*t;
                 txSetColor     (RGB (255, 0, 0), 5);
